% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{.active_element}
\alias{.active_element}
\title{Read the active_element from root input}
\usage{
.active_element(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A session object. Probably always use the default.}
}
\value{
The value of \code{shinyfocuspkg-active_element} as a character scalar (or
NULL).
}
\description{
Read the active_element from root input
}
\keyword{internal}
