% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe.R
\name{.previous_element}
\alias{.previous_element}
\title{Read the previous_element from root input}
\usage{
.previous_element(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A session object. Probably always use the default.}
}
\value{
The value of \code{shinyfocuspkg-previous_element} as a character scalar
(or NULL).
}
\description{
Read the previous_element from root input
}
\keyword{internal}
