% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{shinyfocus_js_dependency}
\alias{shinyfocus_js_dependency}
\title{Add the shinyfocus JavaScript to shiny}
\usage{
shinyfocus_js_dependency()
}
\value{
An \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}, which shiny uses to add the
\code{shinyfocus.js} script exactly once.
}
\description{
Add the \code{shinyfocus.js} script to a shiny app exactly once. Adding this
script setups up the object used to detect changes in focus.
}
\examples{
shinyfocus_js_dependency()
}
